requirements.txt
streamlit==1.26.0
SpeechRecognition==3.14.1
google-generativeai==0.8.3
PyMuPDF